+++ test/components/popover_component_test.rb
@@ -22,14 +22,14 @@ class PrimerPopoverComponentTest < Minitest::Test
 
   def test_allows_customization
     render_inline(Primer::PopoverComponent.new(
-                    position: :absolute, tag: :span, classes: "custom-class"
+                    position: :absolute, classes: "custom-class"
                   )) do |component|
       component.body do
         "Hi there"
       end
     end
 
-    assert_selector("span.Popover.position-absolute.custom-class", text: "Hi there")
+    assert_selector(".Popover.position-absolute.custom-class", text: "Hi there")
   end
 
   def test_respects_message_caret_option