+++ test/components/tooltip_test.rb
@@ -26,6 +26,13 @@ class PrimerTooltipTest < Minitest::Test
     assert_selector(".tooltipped.tooltipped-align-right-1")
   end
 
+  def test_falls_back_when_tag_isnt_valid
+    without_fetch_or_fallback_raises do
+      render_inline(Primer::Tooltip.new(tag: :h5, label: "More")) { "Some" }
+      assert_selector("span.tooltipped")
+    end
+  end
+
   def test_renders_with_no_delay
     render_inline(Primer::Tooltip.new(label: "More", no_delay: true)) { "Some" }